Move chainopts into eth_chainservice
